tests/wpt/tests/css/css-text/parsing/text-group-align-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-group-align with valid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-group-align-property">
<meta name="assert" content="text-group-align supports the full grammar 'none | start | end | left | right | center'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-group-align", "none");
test_valid_value("text-group-align", "start");
test_valid_value("text-group-align", "end");
test_valid_value("text-group-align", "left");
test_valid_value("text-group-align", "right");
test_valid_value("text-group-align", "center");
test_valid_value("text-group-align", "initial");
test_valid_value("text-group-align", "inherit");
test_valid_value("text-group-align", "unset");
test_valid_value("text-group-align", "revert");
test_valid_value("text-group-align", "revert-layer");
</script>
</body>
</html>
